admin/requirements.txt
psycopg
psycopg2-binary
pyyaml
